🌋👠 Checkpoint ./events/message.js:1076878/21 ./commands/math.js:1076878/105 ./commands/report.js:1076878/636

+++ events/message.js
@@ -135,7 +135,7 @@ exports.run = (client, args) => {
         }
       }
     }
-    if (msg.channel.topic.includes("no-commands")) {
+    if (msg.channel.topic && msg.channel.topic.includes("no-commands")) {
       if (msg.deletable()) msg.delete();
       return;
     }
